Moved error unwrapping into common location.
Now the installer and state tool can share common error handling/reporting code.
